Move error page, mini list and tag rendering to templates/misc.dart
