+++ app/lib/frontend/handlers/admin.dart
@@ -12,8 +12,8 @@ import '../../shared/handlers.dart';
 
 import '../backend.dart';
 import '../models.dart';
-import '../templates.dart';
 import '../templates/admin.dart';
+import '../templates/misc.dart';
 
 /// Handles requests for /authorized
 shelf.Response authorizedHandler(_) => htmlResponse(renderAuthorizedPage());
@@ -60,7 +60,7 @@ Future<shelf.Response> adminConfirmHandler(shelf.Request request) async {
 Future<shelf.Response> _formattedInviteExpiredHandler(shelf.Request request,
     [String message = '']) async {
   return htmlResponse(
-    templateService.renderErrorPage(
+    renderErrorPage(
         'Invite expired',
         'The URL you have clicked expired or became invalid.\n\n$message\n',
         null),
