+++ app/lib/frontend/templates.dart
@@ -26,6 +26,7 @@ import 'template_consts.dart';
 import 'templates/_cache.dart';
 import 'templates/_utils.dart';
 import 'templates/layout.dart';
+import 'templates/misc.dart';
 
 /// [TemplateService] singleton instance.
 /// TODO: remove after https://github.com/dart-lang/pub-dartlang-dart/issues/1907 gets fixed
@@ -130,7 +131,7 @@ class TemplateService {
         'dev_version_url': urls.pkgPageUrl(view.name, version: view.devVersion),
         'last_uploaded': view.shortUpdated,
         'desc': view.ellipsizedDescription,
-        'tags_html': _renderTags(
+        'tags_html': renderTags(
           view.platforms,
           isAwaiting: view.isAwaiting,
           isDiscontinued: view.isDiscontinued,
@@ -435,7 +436,7 @@ class TemplateService {
               urls.pkgPageUrl(package.name, version: latestDevVersion.version),
           'dev_name': latestDevVersion.version,
         },
-        'tags_html': _renderTags(
+        'tags_html': renderTags(
           analysis?.platforms,
           isAwaiting: isAwaiting,
           isDiscontinued: card?.isDiscontinued ?? false,
@@ -657,22 +658,6 @@ class TemplateService {
   }
 
   /// Renders the `views/index.mustache` template.
-  String renderErrorPage(
-      String title, String message, List<PackageView> topPackages) {
-    final hasTopPackages = topPackages != null && topPackages.isNotEmpty;
-    final topPackagesHtml = hasTopPackages ? renderMiniList(topPackages) : null;
-    final values = {
-      'title': title,
-      'message_html': markdownToHtml(message, null),
-      'has_top_packages': hasTopPackages,
-      'top_packages_html': topPackagesHtml,
-    };
-    final String content = templateCache.renderTemplate('error', values);
-    return renderLayoutPage(PageType.package, content,
-        title: title, includeSurvey: false);
-  }
-
-  /// Renders the `views/index.mustache` template.
   String renderIndexPage(
     String topHtml,
     String platform,
@@ -701,28 +686,6 @@ class TemplateService {
     );
   }
 
-  /// Renders the `views/mini_list.mustache` template.
-  String renderMiniList(List<PackageView> packages) {
-    final values = {
-      'packages': packages.map((package) {
-        return {
-          'name': package.name,
-          'package_url': urls.pkgPageUrl(package.name),
-          'ellipsized_description': package.ellipsizedDescription,
-          'tags_html': _renderTags(
-            package.platforms,
-            isAwaiting: package.isAwaiting,
-            isDiscontinued: package.isDiscontinued,
-            isLegacy: package.isLegacy,
-            isObsolete: package.isObsolete,
-            packageName: package.name,
-          ),
-        };
-      }).toList(),
-    };
-    return templateCache.renderTemplate('mini_list', values);
-  }
-
   /// Renders the `views/pagination.mustache` template.
   String renderPagination(PageLinks pageLinks) {
     final values = {
@@ -730,60 +693,6 @@ class TemplateService {
     };
     return templateCache.renderTemplate('pagination', values,
         escapeValues: false);
-  }
-
-  /// Renders the tags using the pkg/tags template.
-  String _renderTags(
-    List<String> platforms, {
-    @required bool isAwaiting,
-    @required bool isDiscontinued,
-    @required bool isLegacy,
-    @required bool isObsolete,
-    String packageName,
-  }) {
-    final List<Map> tags = <Map>[];
-    if (isAwaiting) {
-      tags.add({
-        'status': 'missing',
-        'text': '[awaiting]',
-        'title': 'Analysis should be ready soon.',
-      });
-    } else if (isDiscontinued) {
-      tags.add({
-        'status': 'discontinued',
-        'text': '[discontinued]',
-        'title': 'Package was discontinued.',
-      });
-    } else if (isObsolete) {
-      tags.add({
-        'status': 'missing',
-        'text': '[outdated]',
-        'title': 'Package version too old, check latest stable.',
-      });
-    } else if (isLegacy) {
-      tags.add({
-        'status': 'legacy',
-        'text': 'Dart 2 incompatible',
-        'title': 'Package does not support Dart 2.',
-      });
-    } else if (platforms != null && platforms.isNotEmpty) {
-      tags.addAll(platforms.map((platform) {
-        final platformDict = getPlatformDict(platform, nullIfMissing: true);
-        return {
-          'text': platformDict?.name ?? platform,
-          'href': platformDict?.listingUrl,
-          'title': platformDict?.tagTitle,
-        };
-      }));
-    } else {
-      tags.add({
-        'status': 'unidentified',
-        'text': '[unidentified]',
-        'title': 'Check the analysis tab for further details.',
-        'href': urls.analysisTabUrl(packageName),
-      });
-    }
-    return templateCache.renderTemplate('pkg/tags', {'tags': tags});
   }
 }
