+++ app/test/frontend/templates_test.dart
@@ -19,6 +19,7 @@ import 'package:pub_dartlang_org/frontend/static_files.dart';
 import 'package:pub_dartlang_org/frontend/templates.dart';
 import 'package:pub_dartlang_org/frontend/templates/admin.dart';
 import 'package:pub_dartlang_org/frontend/templates/layout.dart';
+import 'package:pub_dartlang_org/frontend/templates/misc.dart';
 
 import '../shared/html_validation.dart';
 import 'utils.dart';
@@ -65,7 +66,7 @@ void main() {
     }
 
     test('index page', () {
-      final popularHtml = templateService.renderMiniList([
+      final popularHtml = renderMiniList([
         new PackageView.fromModel(
           package: testPackage,
           version: testPackageVersion,
@@ -80,7 +81,7 @@ void main() {
     });
 
     test('landing page flutter', () {
-      final popularHtml = templateService.renderMiniList([
+      final popularHtml = renderMiniList([
         new PackageView.fromModel(
           package: testPackage,
           version: testPackageVersion,
@@ -96,7 +97,7 @@ void main() {
     });
 
     test('landing page web', () {
-      final popularHtml = templateService.renderMiniList([
+      final popularHtml = renderMiniList([
         new PackageView.fromModel(
           package: testPackage,
           version: testPackageVersion,
@@ -465,8 +466,7 @@ void main() {
     });
 
     test('error page', () {
-      final String html =
-          templateService.renderErrorPage('error_title', 'error_message', [
+      final String html = renderErrorPage('error_title', 'error_message', [
         new PackageView(
           name: 'popular_pkg',
           version: '1.0.2',
